backend/app/automation/web_scraper.py
from playwright.async_api import async_playwright, Browser, Page, TimeoutError as PlaywrightTimeoutError
from bs4 import BeautifulSoup
import json
import asyncio
from typing import Dict, List, Optional, Any, Tuple
from urllib.parse import urljoin, urlparse
import re
from dataclasses import dataclass
from app.utils.logging import get_logger, log_automation_event

logger = get_logger(__name__)


@dataclass
class FormField:
    """Represents a form field found during scraping."""
    name: str
    type: str
    selector: str
    required: bool
    placeholder: str = ""
    label: str = ""
    validation_pattern: str = ""
    options: List[str] = None  # For select/radio fields


@dataclass
class SignupFormAnalysis:
    """Analysis results of a signup form."""
    form_selector: str
    action_url: str
    method: str
    fields: List[FormField]
    submit_button_selector: str
    has_captcha: bool = False
    has_terms_checkbox: bool = False
    requires_email_verification: bool = False
    additional_steps: List[str] = None


class WebScraper:
    """Web scraper for analyzing signup processes."""
    
    def __init__(self, headless: bool = True, timeout: int = 30000):
        self.headless = headless
        self.timeout = timeout
        self.browser: Optional[Browser] = None
        self.page: Optional[Page] = None
    
    async def start(self):
        """Start the browser instance."""
        try:
            playwright = await async_playwright().start()
            self.browser = await playwright.chromium.launch(headless=self.headless)
            self.page = await self.browser.new_page()
            logger.info("Web scraper browser started")
        except Exception as e:
            logger.error(f"Failed to start browser: {str(e)}")
            raise
    
    async def close(self):
        """Close the browser instance."""
        try:
            if self.browser:
                await self.browser.close()
                logger.info("Web scraper browser closed")
        except Exception as e:
            logger.error(f"Error closing browser: {str(e)}")
    
    async def analyze_signup_page(self, url: str) -> SignupFormAnalysis:
        """Analyze a signup page to understand its structure."""
        if not self.page:
            raise RuntimeError("Browser not started")
        
        try:
            log_automation_event("page_analysis_start", {"url": url})
            
            # Navigate to the page
            await self.page.goto(url, wait_until="networkidle", timeout=self.timeout)
            
            # Basic form analysis (simplified for now)
            form_analysis = SignupFormAnalysis(
                form_selector="form",
                action_url="/signup",
                method="post",
                fields=[
                    FormField(name="email", type="email", selector="input[type='email']", required=True),
                    FormField(name="password", type="password", selector="input[type='password']", required=True)
                ],
                submit_button_selector="button[type='submit']"
            )
            
            log_automation_event("page_analysis_complete", {"url": url})
            return form_analysis
            
        except Exception as e:
            log_automation_event("page_analysis_error", {"url": url, "error": str(e)})
            logger.error(f"Error analyzing signup page {url}: {str(e)}")
            raise


async def analyze_website_signup(url: str) -> SignupFormAnalysis:
    """Analyze a website's signup process."""
    scraper = WebScraper()
    try:
        await scraper.start()
        return await scraper.analyze_signup_page(url)
    finally:
        await scraper.close() 