backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Security & Encryption
cryptography>=3.4.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Web Automation & Scraping
playwright==1.40.0
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
lxml==4.9.3

# AI & Natural Language Processing
openai==1.3.8
langchain==0.0.335
tiktoken==0.5.2

# Data Processing
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
rich==13.7.0
loguru==0.7.2
httpx==0.25.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1 